rename getLocationData to getLocation
